requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
slack-sdk==3.33.0
python-dotenv>=1.0.1,<2
gspread>=6.0.0,<7
oauth2client>=4.1.3,<5
requests>=2.28.0,<3
pydantic>=2.0.0,<3
